Created Public Function to Access Plugin State
I created a function that will check if the plugin state is none. After
the page has loaded, this will reliably show that no plugin has been
installed.

If the plugin has been installed, the state will be injecting /
injected.

The current function, Temasys.WebRTCPlugin.isPluginInstalled, takes too
many options and does not have a return value.

It would be better not to access Temasys.WebRTCPlugin.pluginState in my
external code, as this internal API could change in the future.